% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_depth_distribution.R
\name{calculate_tree_depth}
\alias{calculate_tree_depth}
\title{Calculate depth in a tree}
\usage{
calculate_tree_depth(frame)
}
\arguments{
\item{frame}{A data frame returned by the function getTree() applied to a random forest}
}
\value{
The original data frame with an additional column \code{depth}
}
\description{
Calculate the depth of each node in a tree
}
\examples{
calculate_tree_depth(getTree(randomForest(Species ~ ., data = iris), k = 1, labelVar = T))

}
