% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_depth_interactions.R
\name{conditional_depth}
\alias{conditional_depth}
\title{Calculate conditional depth in a tree}
\usage{
conditional_depth(frame, vars)
}
\arguments{
\item{frame}{A data frame returned by the function calculate_tree_depth(getTree()) applied to a random forest}

\item{vars}{A character vector with variables with respect to which conditional minimal depth will be calculated}
}
\value{
The original data frame with an additional column \code{depth}
}
\description{
For a vactor of conditioning variables insert values of conditional depth to a data frame of proper structure
}
\examples{
conditional_depth(calculate_tree_depth(getTree(randomForest(Species ~ ., data = iris), k = 1, labelVar = T)), vars = names(iris))

}
