% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_importance.R
\name{measure_importance}
\alias{measure_importance}
\title{Get a data frame with various measures of importance of variables in a forest}
\usage{
measure_importance(forest)
}
\arguments{
\item{forest}{A random forest produced by the function randomForest with option importance = TRUE}
}
\value{
A data frame with rows corresponding to variables and columns to various measures of importance of variables
}
\description{
Get a data frame with various measures of importance of variables in a forest
}
\examples{
measure_importance(randomForest(Species ~ ., data = iris))
}
