% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_depth_distribution.R
\name{min_depth_count}
\alias{min_depth_count}
\title{Count the trees in which each variable had a given minimal depth}
\usage{
min_depth_count(min_depth_frame)
}
\arguments{
\item{min_depth_frame}{A data frame output of min_depth_distribution or min_depth_distribution_memory function}
}
\value{
A data frame with count of occurances of each minimal depth value including NA's
}
\description{
Count the trees in which each variable had a given minimal depth
}
\examples{
min_depth_count(min_depth_distribution(randomForest(Species ~ ., data = iris)))
}
