% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_depth_distribution.R
\name{min_depth_distribution}
\alias{min_depth_distribution}
\title{Get minimal depth values for all trees in a random forest}
\usage{
min_depth_distribution(forest)
}
\arguments{
\item{forest}{A randomForest object}
}
\value{
A data frame with the value of minimal depth for every variable in every tree
}
\description{
Get minimal depth values for all trees in a random forest
}
\examples{
min_depth_distribution(randomForest(Species ~ ., data = iris))
}
