% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_depth_interactions.R
\name{min_depth_interactions_means}
\alias{min_depth_interactions_means}
\title{Count the trees in which each variable had a given conditional minimal depth}
\usage{
min_depth_interactions_means(forest, vars)
}
\arguments{
\item{forest}{A randomForest object}

\item{vars}{A character vector with variables with respect to which conditional minimal depth will be calculated}
}
\value{
A list with means of conditional minimal depth in the first element and the size of sample for each mean in the second element
}
\description{
Count the trees in which each variable had a given conditional minimal depth
}
\examples{
min_depth_interactions_means(randomForest(Species ~ ., data = iris), vars = names(iris))
}
