% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_depth_interactions.R
\name{min_depth_interactions_values}
\alias{min_depth_interactions_values}
\title{Get a data frame with values of minimal depth conditional on selected variables}
\usage{
min_depth_interactions_values(forest, vars)
}
\arguments{
\item{forest}{A randomForest object}

\item{vars}{A character vector with variables with respect to which conditional minimal depth will be calculated}
}
\value{
A data frame with the value of minimal depth conditional on variables from the supplied vector
}
\description{
Get a data frame with values of minimal depth conditional on selected variables
}
\examples{
min_depth_interactions_values(randomForest(Species ~ ., data = iris), vars = names(iris))
}
