% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_importance.R
\name{plot_importance_ggpairs}
\alias{plot_importance_ggpairs}
\title{Plot selected measures of importance of variables in a forest using ggpairs}
\usage{
plot_importance_ggpairs(importance_frame,
  measures = names(importance_frame)[c(2, 4, 5, 6)],
  main = "Relations between measures of importance")
}
\arguments{
\item{importance_frame}{A result of using the function measure_importance() to a random forest}

\item{measures}{A character vector specifying the measures of importance to be used}

\item{main}{A string to be used as title of the plot}
}
\value{
A ggplot object
}
\description{
Plot selected measures of importance of variables in a forest using ggpairs
}
\examples{
plot_importance_ggpairs(measure_importance(randomForest(Species ~ ., data = iris)))
}
