% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_importance.R
\name{plot_multi_way_importance}
\alias{plot_multi_way_importance}
\title{Plot two or three measures of importance of variables in a forest}
\usage{
plot_multi_way_importance(importance_frame, x_measure = "mean_minimal_depth",
  y_measure = "times_a_root", size_measure = NULL, min_no_of_trees = 0.1 *
  max(importance_frame$no_of_trees), x_label_prob = 0.01,
  y_label_prob = 0.01, size_label_prob = 0.01,
  main = "Multi-way importance plot")
}
\arguments{
\item{importance_frame}{A result of using the function measure_importance() to a random forest}

\item{x_measure}{The measure of importance to be shown on the X axis}

\item{y_measure}{The measure of importance to be shown on the Y axis}

\item{size_measure}{The measure of importance to be shown as size of points (optional)}

\item{min_no_of_trees}{The minimal number of trees in which a variable has to be used for splitting to be used for plotting}

\item{x_label_prob}{The fraction of best points according to the measure on X axis to be labeled}

\item{y_label_prob}{The fraction of best points according to the measure on Y axis to be labeled}

\item{size_label_prob}{The fraction of best points according to the measure displayed by size of points to be labeled}

\item{main}{A string to be used as title of the plot}
}
\value{
A ggplot object
}
\description{
Plot two or three measures of importance of variables in a forest
}
\examples{
plot_multi_way_importance(measure_importance(randomForest(Species ~ ., data = iris)))
}
