% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_forest.R
\name{summarize_forest}
\alias{summarize_forest}
\title{Summarize a random forest}
\usage{
summarize_forest(forest, interactions = FALSE, vars = NULL, measures = if
  (forest$type == "classification") c("mean_minimal_depth", "accuracy_decrease",
  "gini_decrease", "no_of_nodes", "times_a_root") else c("mean_minimal_depth",
  "mse_increase", "node_purity_increase", "no_of_nodes", "times_a_root"))
}
\arguments{
\item{forest}{A randomForest object created with the options importance = TRUE and localImp = TRUE}

\item{interactions}{Logical value: should variable interactions be considered (this may be time-consuming)}

\item{vars}{A character vector with variables with respect to which interactions will be considered if NULL then they will be selected using the important_variables() function}

\item{measures}{A character vector specifying the importance measures to be used for plotting ggpairs}
}
\value{
A html document in your working directory
}
\description{
Summarizes a random forest in a html document using plots created by randomForestExplainer
}
\examples{
summarize_forest(randomForest(Species ~ ., data = iris, importance = TRUE, localImp = TRUE), vars = names(iris), interactions = TRUE)

}
