% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_importance.R
\name{measure_importance}
\alias{measure_importance}
\title{Importance of variables in a random forest}
\usage{
measure_importance(forest)
}
\arguments{
\item{forest}{A random forest produced by the function randomForest with option localImp = TRUE}
}
\value{
A data frame with rows corresponding to variables and columns to various measures of importance of variables
}
\description{
Get a data frame with various measures of importance of variables in a random forest
}
\examples{
measure_importance(randomForest::randomForest(Species ~ ., data = iris, localImp = TRUE))

}
